refactor out into supporting library for si client
