+++ ci/commands/manage-image-lifecycle.ts
@@ -4,186 +4,7 @@
  */
 
 import { execSync } from 'child_process';
-
-class SystemInitiativeClient {
-  private apiToken: string;
-  private workspaceId: string;
-  private apiUrl = "https://api.systeminit.com";
-
-  constructor() {
-    this.apiToken = process.env.SI_API_TOKEN || "";
-    this.workspaceId = process.env.SI_WORKSPACE_ID || "";
-
-    if (!this.apiToken || !this.workspaceId) {
-      throw new Error(
-        "Missing required environment variables: SI_API_TOKEN or SI_WORKSPACE_ID"
-      );
-    }
-  }
-
-  private get headers() {
-    return {
-      "Authorization": `Bearer ${this.apiToken}`,
-      "Content-Type": "application/json",
-    };
-  }
-
-  private async fetch(
-    url: string,
-    options: RequestInit = {}
-  ): Promise<Response> {
-    const response = await fetch(url, {
-      ...options,
-      headers: {
-        ...this.headers,
-        ...options.headers,
-      },
-    });
-    return response;
-  }
-
-  async createChangeSet(name: string): Promise<any> {
-    try {
-      const response = await this.fetch(
-        `${this.apiUrl}/v1/w/${this.workspaceId}/change-sets`,
-        {
-          method: "POST",
-          body: JSON.stringify({ changeSetName: name }),
-        }
-      );
-
-      if (!response.ok) {
-        const errorText = await response.text();
-        throw new Error(
-          `HTTP ${response.status}: ${response.statusText} - ${errorText}`
-        );
-      }
-
-      return await response.json();
-    } catch (e) {
-      throw new Error(`Failed to create change set '${name}': ${e}`);
-    }
-  }
-
-  async findComponentByName(changeSetId: string, componentName: string): Promise<any> {
-    try {
-      const response = await this.fetch(
-        `${this.apiUrl}/v1/w/${this.workspaceId}/change-sets/${changeSetId}/components/find?component=${encodeURIComponent(componentName)}`
-      );
-
-      if (!response.ok) {
-        const errorText = await response.text();
-        throw new Error(
-          `HTTP ${response.status}: ${response.statusText} - ${errorText}`
-        );
-      }
-
-      const data = await response.json() as any;
-      return data.component || null;
-    } catch (e) {
-      throw new Error(`Failed to find component '${componentName}': ${e}`);
-    }
-  }
-
-  async updateComponentAttribute(
-    changeSetId: string,
-    componentId: string,
-    attributePath: string,
-    value: any
-  ): Promise<any> {
-    try {
-      const response = await this.fetch(
-        `${this.apiUrl}/v1/w/${this.workspaceId}/change-sets/${changeSetId}/components/${componentId}`,
-        {
-          method: "PUT",
-          body: JSON.stringify({
-            attributes: {
-              [attributePath]: value
-            }
-          }),
-        }
-      );
-
-      if (!response.ok) {
-        const errorText = await response.text();
-        throw new Error(
-          `HTTP ${response.status}: ${response.statusText} - ${errorText}`
-        );
-      }
-
-      return await response.json();
-    } catch (e) {
-      throw new Error(`Failed to update component attribute '${attributePath}': ${e}`);
-    }
-  }
-
-  async applyChangeSet(changeSetId: string, timeoutSeconds = 120, retryInterval = 5): Promise<any> {
-    const startTime = Date.now();
-
-    while (Date.now() - startTime < timeoutSeconds * 1000) {
-      try {
-        const response = await this.fetch(
-          `${this.apiUrl}/v1/w/${this.workspaceId}/change-sets/${changeSetId}/force_apply`,
-          {
-            method: "POST",
-            body: "",
-          }
-        );
-
-        if (!response.ok) {
-          if (response.status === 428) {
-            const elapsed = Date.now() - startTime;
-            const remaining = timeoutSeconds * 1000 - elapsed;
-            console.log(
-              `⏳ Dependent values still calculating. Retrying in ${retryInterval}s... (${
-                (remaining / 1000).toFixed(1)
-              }s remaining)`
-            );
-
-            if (remaining > retryInterval * 1000) {
-              await new Promise((resolve) =>
-                setTimeout(resolve, retryInterval * 1000)
-              );
-              continue;
-            } else {
-              break;
-            }
-          } else {
-            const errorText = await response.text();
-            throw new Error(
-              `HTTP ${response.status}: ${response.statusText} - ${errorText}`
-            );
-          }
-        }
-
-        console.log("✅ Change set applied successfully");
-        return await response.json();
-      } catch (e) {
-        if (e instanceof Error && e.message.includes("428")) {
-          continue;
-        }
-        throw e;
-      }
-    }
-
-    throw new Error(
-      `❌ Change set apply timeout: Failed to apply after ${timeoutSeconds}s - dependent values still calculating`
-    );
-  }
-
-  async deleteChangeSet(changeSetId: string): Promise<any> {
-    const response = await this.fetch(
-      `${this.apiUrl}/v1/w/${this.workspaceId}/change-sets/${changeSetId}`,
-      { method: "DELETE" }
-    );
-
-    if (!response.ok) {
-      throw new Error(`HTTP ${response.status}: ${response.statusText}`);
-    }
-
-    return await response.json();
-  }
-}
+import { SystemInitiativeClient } from '../lib/system-initiative-client.js';
 
 export type ImageAction = 'build' | 'publish' | 'push' | 'deploy';
 export type Component = 'api' | 'web' | 'e2e';
@@ -378,7 +199,7 @@ class ImageLifecycleManager {
       // Step 4: Apply the change set
       console.log(`📤 Applying change set ${changeSetId}...`);
       
-      await siClient.applyChangeSet(changeSetId);
+      await siClient.forceApplyChangeSet(changeSetId);
       
       console.log(`✅ Change set applied successfully`);
       console.log(`🎉 Deployment complete for ${component.toUpperCase()}`);
